ansible/roles/pipecatapp/files/requirements.txt
pipecat-ai
RealtimeSTT
faster-whisper
MeloTTS
torch
torchaudio
ultralytics
opencv-python-headless
kittentts
soundfile
sentence-transformers
faiss-cpu
paramiko
